refactor: User repository 패키지 이동
