+++ src/main/java/woowa/camp/jspcafe/configuration/AppContextListener.java
@@ -8,8 +8,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import woowa.camp.jspcafe.repository.article.ArticleRepository;
 import woowa.camp.jspcafe.repository.article.InMemoryArticleRepository;
-import woowa.camp.jspcafe.repository.InMemoryUserRepository;
-import woowa.camp.jspcafe.repository.UserRepository;
+import woowa.camp.jspcafe.repository.user.InMemoryUserRepository;
+import woowa.camp.jspcafe.repository.user.UserRepository;
 import woowa.camp.jspcafe.service.ArticleService;
 import woowa.camp.jspcafe.service.UserService;
 import woowa.camp.jspcafe.utils.time.DateTimeProvider;
