+++ src/test/java/woowa/camp/jspcafe/service/ArticleServiceTest.java
@@ -16,8 +16,8 @@ import woowa.camp.jspcafe.fixture.ArticleFixture;
 import woowa.camp.jspcafe.fixture.UserFixture;
 import woowa.camp.jspcafe.repository.article.ArticleRepository;
 import woowa.camp.jspcafe.repository.article.InMemoryArticleRepository;
-import woowa.camp.jspcafe.repository.InMemoryUserRepository;
-import woowa.camp.jspcafe.repository.UserRepository;
+import woowa.camp.jspcafe.repository.user.InMemoryUserRepository;
+import woowa.camp.jspcafe.repository.user.UserRepository;
 import woowa.camp.jspcafe.service.dto.ArticleDetailsResponse;
 import woowa.camp.jspcafe.service.dto.ArticlePreviewResponse;
 import woowa.camp.jspcafe.service.dto.ArticleWriteRequest;
