+++ src/test/java/woowa/camp/jspcafe/service/UserServiceTest.java
@@ -14,8 +14,8 @@ import org.junit.jupiter.params.provider.ValueSource;
 import woowa.camp.jspcafe.domain.User;
 import woowa.camp.jspcafe.domain.exception.UserException;
 import woowa.camp.jspcafe.fixture.UserFixture;
-import woowa.camp.jspcafe.repository.InMemoryUserRepository;
-import woowa.camp.jspcafe.repository.UserRepository;
+import woowa.camp.jspcafe.repository.user.InMemoryUserRepository;
+import woowa.camp.jspcafe.repository.user.UserRepository;
 import woowa.camp.jspcafe.service.dto.RegistrationRequest;
 import woowa.camp.jspcafe.service.dto.UserResponse;
 import woowa.camp.jspcafe.utils.FixedDateTimeProvider;
